+++ packages/vector_graphics/CHANGELOG.md
@@ -12,6 +12,7 @@
 * Adds SVG `feBlend` support with parameter, pixel, and visual reference tests.
 * Adds SVG `feComposite` support with parameter, pixel, and visual reference tests.
 * Adds SVG `feDropShadow` support with parameter, pixel, and visual reference tests.
+* Adds SVG `feMorphology` support with parameter, pixel, and visual reference tests.
 
 ## 1.2.3